# fms-ui/src/assets/fms-api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fleet Management System API Documentation",
    "description": "Dokumentacja API - autoryzacja za pomocą JWT",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/admin/changeUserData/{userId}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "changeUserData",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/createUser": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedUserCredentialsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/resetPasswordForUser/{userId}": {
      "patch": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "resetPasswordForUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedUserCredentialsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/changeUserRole/{userId}": {
      "patch": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "changeUserRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/deleteUser/{userId}": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/deleteService/{serviceId}": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteService",
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/deleteRefueling/{refuelingId}": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteRefueling",
        "parameters": [
          {
            "name": "refuelingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/deleteInsurance/{insuranceId}": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteInsurance",
        "parameters": [
          {
            "name": "insuranceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/deleteInspection/{inspectionId}": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteInspection",
        "parameters": [
          {
            "name": "inspectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/admin/deleteAssignment/{assignmentId}": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssignmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "vehicleId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ChangeUserDataRequest": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "DRIVER",
              "IOT"
            ]
          },
          "assignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssignmentDto"
            }
          }
        }
      },
...
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "DRIVER",
              "IOT"
            ]
          }
        }
      },
      "GeneratedUserCredentialsResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "generatedPassword": {
            "type": "string"
          }
        }
      },
...
      "ChangeUserRoleRequest": {
        "type": "object",
        "properties": {
          "newRole": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "DRIVER",
              "IOT"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}